training/requirements.txt
# Training Pipeline - Python Dependencies
# Install: pip install -r requirements.txt

# YOLOv8
ultralytics>=8.2.0

# PyTorch (installed as ultralytics dependency, but pinning for clarity)
torch>=2.0.0
torchvision>=0.15.0

# Image processing
opencv-python>=4.8.0
pillow>=11.0.0

# ML utilities
scikit-learn>=1.5.0
scikit-image>=0.22.0
numpy>=2.0.0

# Visualization
matplotlib>=3.8.0

# Progress bars
tqdm>=4.66.0
